[CHORE] CreateLetterViewController 내부에 있는 ui component 코드 CreateLetterView로 이동(#391)
